fix(approve): stop tripping Vercel Hobby cron limits, send inline via after()
Vercel Hobby caps cron jobs at once-per-day — `* * * * *` was killing the
deploy. Flipping the model so the cron is just a safety net:

- Extracted drainOutboxBatch() into lib/email/drain-outbox.ts
- Server Actions (submitDocument, submitSignature) call after() to drain
  the outbox the moment the response is sent — users see mail within seconds,
  no Vercel plan upgrade required
- Cron now runs 04:00 daily as a retry sweep for anything after() missed
- Cron route is a thin wrapper around the shared drain logic
