+++ vercel.json
@@ -3,7 +3,7 @@
   "crons": [
     {
       "path": "/api/cron/approve-emails",
-      "schedule": "* * * * *"
+      "schedule": "0 4 * * *"
     }
   ]
 }
